safe rotation in settings menu, reset game icon
